Highlight block when stepping.
